# Application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music library API",
        "description": "Music library API",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/audio/play/{path}": {
            "get": {
                "tags": [
                    "Audio"
                ],
                "summary": "Play an audio by pathname",
                "description": "Play an audio by pathname",
                "operationId": "acf101771b0bbe2a3b5400eedb0e550d",
                "parameters": [
                    {
                        "name": "path",
                        "in": "path",
                        "description": "Path of the file(post audiofile)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The audio file content",
                        "content": {
                            "audio/mpeg": {}
                        }
                    },
                    "404": {
                        "description": "Audiofile not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}